test: 우승자 구하는 테스트 작성
